+++ dbdaora/hash/_tests/datastore/test_integration_service_hash_aioredis_datastore_add.py
@@ -1,4 +1,3 @@
-import asynctest
 import pytest
 from aioredis import RedisError
 
@@ -16,9 +15,9 @@ async def test_should_add(fake_service, fake_entity):
 async def test_should_add_to_fallback_after_open_circuit_breaker(
     fake_service, fake_entity, mocker
 ):
-    fake_service.repository.memory_data_source.hmset = asynctest.CoroutineMock(
-        side_effect=RedisError
-    )
+    fake_multi_exec = mocker.MagicMock()
+    fake_multi_exec.return_value.execute.side_effect = RedisError
+    fake_service.repository.memory_data_source.multi_exec = fake_multi_exec
     await fake_service.add(fake_entity)
 
     entity = await fake_service.get_one('fake', other_id='other_fake')